# backend/docs/swagger.yaml
definitions:
...
  domain.ProductBundle:
    properties:
      bundle_id:
        type: integer
      layer_no:
        type: integer
      product_id:
        type: integer
    type: object
...
  domain.TestRank:
    properties:
      distance_behind:
        type: integer
      is_public:
        type: boolean
      product_id:
        type: integer
      rank:
        type: integer
      test_id:
        type: integer
      version:
        type: string
    type: object
...
  rankingsHandler.RankingsPOSTRequest:
    properties:
      is_public:
        type: boolean
      rank:
        type: integer
      test_id:
        type: integer
      version:
        type: string
      wins:
        type: integer
    required:
    - rank
    - test_id
    - wins
    type: object
...
  userProfileHandler.TeamResponse:
    properties:
      name:
        type: string
      team_role:
        type: string
    type: object
  userProfileHandler.UserProfileResponse:
    properties:
      email:
        type: string
      team:
        $ref: '#/definitions/userProfileHandler.TeamResponse'
      user_role:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /bundles:
    get:
      consumes:
      - application/json
      description: Retrieves a list of bundles based on query parameters.
      parameters:
      - description: Bundle ID
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successful response with a list of bundles
          schema:
            items:
              $ref: '#/definitions/domain.ProductBundle'
            type: array
        "500":
          description: Could not retrieve all bundles.
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get a list of bundles
      tags:
      - Bundles
    post:
      consumes:
      - application/json
      description: Creates a new bundle based on the provided JSON request body.
      parameters:
      - description: Bundle details
        in: body
        name: bundle
        required: true
        schema:
          $ref: '#/definitions/domain.ProductBundle'
      produces:
      - application/json
      responses:
        "201":
          description: Bundle created successfully
          schema:
            type: string
        "400":
          description: Invalid request body
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Failed to commit the transaction.
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new bundle
      tags:
      - Bundles
  /bundles/:
    get:
      consumes:
      - application/json
      description: Retrieves a list of bundles based on query parameters.
      parameters:
      - description: Bundle ID
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successful response with a list of bundles
          schema:
            items:
              $ref: '#/definitions/domain.ProductBundle'
            type: array
        "500":
          description: Could not retrieve all bundles.
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get a list of bundles
      tags:
      - Bundles
    post:
      consumes:
      - application/json
      description: Creates a new bundle based on the provided JSON request body.
      parameters:
      - description: Bundle details
        in: body
        name: bundle
        required: true
        schema:
          $ref: '#/definitions/domain.ProductBundle'
      produces:
      - application/json
      responses:
        "201":
          description: Bundle created successfully
          schema:
            type: string
        "400":
          description: Invalid request body
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Failed to commit the transaction.
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new bundle
      tags:
      - Bundles
...
  /rankings:
    get:
      consumes:
      - application/json
      description: Retrieves a list of rankings based on query parameters.
      parameters:
      - description: Public rankings
        in: query
        name: public
        type: string
      - description: Test ID
        in: query
        name: test_id
        type: integer
      - description: Rank
        in: query
        name: rank
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successful response with a list of rankings
          schema:
            items:
              $ref: '#/definitions/domain.TestRank'
            type: array
        "400":
          description: Invalid rank parameter
          schema:
            type: string
        "500":
          description: Could not retrieve all rankings
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get a list of rankings
      tags:
      - Rankings
    post:
      consumes:
      - application/json
      description: Creates a new ranking
      parameters:
      - description: New ranking information
        in: body
        name: ranking
        required: true
        schema:
          $ref: '#/definitions/rankingsHandler.RankingsPOSTRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Ranking created successfully
        "400":
          description: Invalid request body
          schema:
            type: string
        "409":
          description: Product is already part of the test
          schema:
            type: string
        "500":
          description: Could not create ranking
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new ranking
      tags:
      - Rankings
  /rankings/:
    post:
      consumes:
      - application/json
      description: Creates a new ranking
      parameters:
      - description: New ranking information
        in: body
        name: ranking
        required: true
        schema:
          $ref: '#/definitions/rankingsHandler.RankingsPOSTRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Ranking created successfully
        "400":
          description: Invalid request body
          schema:
            type: string
        "409":
          description: Product is already part of the test
          schema:
            type: string
        "500":
          description: Could not create ranking
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new ranking
      tags:
      - Rankings
  /rankings/{rank_id}:
    get:
      consumes:
      - application/json
      description: Retrieves a list of rankings based on query parameters.
      parameters:
      - description: Public rankings
        in: query
        name: public
        type: string
      - description: Test ID
        in: query
        name: test_id
        type: integer
      - description: Rank
        in: query
        name: rank
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successful response with a list of rankings
          schema:
            items:
              $ref: '#/definitions/domain.TestRank'
            type: array
        "400":
          description: Invalid rank parameter
          schema:
            type: string
        "500":
          description: Could not retrieve all rankings
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get a list of rankings
      tags:
      - Rankings
...
  /user/profile:
    get:
      consumes:
      - application/json
      description: Retrieves the user profile for the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: Successful response with user profile details
          schema:
            $ref: '#/definitions/userProfileHandler.UserProfileResponse'
        "400":
          description: Invalid request URL
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: User not found
          schema:
            type: string
        "500":
          description: Could not create user profile struct.
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - UserProfile
...
swagger: "2.0"